requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
boto3>=1.36.0
prisma>=0.15.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
ruff>=0.8.0
bandit>=1.8.0
PyJWT>=2.9.0
google-auth>=2.35.0
